% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_builders.R
\name{build_string_net}
\alias{build_string_net}
\title{Build an inverse-degree & Laplacian matrix from a STRING-PPI sub-network}
\usage{
build_string_net(genes)
}
\arguments{
\item{genes}{character vector of HGNC gene symbols present in STRING.}
}
\value{
A list with elements
\itemize{
\item \strong{A}       – weighted adjacency (\code{dgCMatrix}),
\item \strong{DegInv}  – inverse-degree diagonal (\code{dgCMatrix}),
\item \strong{Lap}     – unnormalised Laplacian (\code{dgCMatrix}),
\item \strong{genes}   – vector of genes kept after mapping.
}
}
\description{
Build an inverse-degree & Laplacian matrix from a STRING-PPI sub-network
}
