% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_builders.R
\name{build_wgcna_net}
\alias{build_wgcna_net}
\title{Construct an unsigned WGCNA adjacency + penalties}
\usage{
build_wgcna_net(expr_mat)
}
\arguments{
\item{expr_mat}{\verb{p × n} expression matrix (genes × samples); \strong{no NAs}.}
}
\value{
Same list structure as \code{\link[=build_string_net]{build_string_net()}}.
}
\description{
Construct an unsigned WGCNA adjacency + penalties
}
