% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{gradient.Weibull}
\alias{gradient.Weibull}
\title{Gradient of the (negative) log-likelihood in a Weibull AFT model}
\usage{
gradient.Weibull(X, Y, eta, delta, sigma, n)
}
\description{
Gradient of the (negative) log-likelihood in a Weibull AFT model
}
\keyword{internal}
