% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_builders.R
\name{isol_status}
\alias{isol_status}
\title{Is a selected gene isolated in the underlying graph?}
\usage{
isol_status(sel_genes, A_mat)
}
\arguments{
\item{sel_genes}{character vector of selected genes.}

\item{A_mat}{sparse adjacency matrix (\code{dgCMatrix}).}
}
\value{
character vector \code{"isolated"} / \code{"connected"} matching \code{sel_genes}.
}
\description{
Is a selected gene isolated in the underlying graph?
}
