% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_helpers.R
\name{metric_fold}
\alias{metric_fold}
\title{Uno-AUC + C-index for one validation fold}
\usage{
metric_fold(time_tr, status_tr, lp_tr, time_va, status_va, lp_va)
}
\description{
Uno-AUC + C-index for one validation fold
}
\keyword{internal}
